v1.0.0

+++ src/lib/Hardware/HardwareManager.ts
@@ -4,9 +4,6 @@ import { getApiURL } from "../Util";
 import { Response, IHardware, HardwareOptions, checkoutOptions, HardwarePostOptions, HardwareUpdateOptions } from "../Interfaces";
 import { Hardware } from "./Hardware";
 
-// Error:
-//		const result = await res.json();
-//		if (await result.status == "error") throw(`Error on checkin:\n${JSON.stringify(result, null, " ")}`);
 export class HardwareManager extends Manager {
 
 	// GET
@@ -26,8 +23,6 @@ export class HardwareManager extends Manager {
 			}
 		});
 		const result = await res.json();
-		if (await result.status == "error") throw (JSON.stringify(result, null, " "));
-		if (res.status !== 200) throw (JSON.stringify(result, null, " "));
 
 		const json: Response<IHardware> = result;
 		return json.rows.map(hardware => new Hardware(hardware));
@@ -39,6 +34,7 @@ export class HardwareManager extends Manager {
 	 * @param id Asset ID
 	 */
 	async getID(id: number) {
+		if(!id) return "id required!";
 		const res = await fetch(getApiURL(this.snipeURL, `/hardware/${id}`), {
 			method: "GET",
 			headers: {
@@ -48,8 +44,6 @@ export class HardwareManager extends Manager {
 			}
 		});
 		const result = await res.json();
-		if (res.status !== 200) throw (JSON.stringify(result, null, " "));
-		if (result.status == "error") throw (JSON.stringify(result, null, " "));
 
 		const json: IHardware = result;
 		return new Hardware(json);
@@ -60,6 +54,7 @@ export class HardwareManager extends Manager {
 	 * @param id Asset ID
 	 */
 	async getAssetTag(asset_tag: string) {
+		if(!asset_tag) return "asset_tag required!";
 		const res = await fetch(getApiURL(this.snipeURL, `/hardware/bytag/${asset_tag}`), {
 			method: "GET",
 			headers: {
@@ -69,8 +64,6 @@ export class HardwareManager extends Manager {
 			}
 		});
 		const result = await res.json();
-		if (res.status !== 200) throw (JSON.stringify(result, null, " "));
-		if (result.status == "error") throw (JSON.stringify(result, null, " "));
 
 		const json: IHardware = result;
 		return new Hardware(json);
@@ -81,6 +74,7 @@ export class HardwareManager extends Manager {
 	 * @param id Asset ID
 	 */
 	async getSerial(serial: string) {
+		if(!serial) return "serial required!";
 		const res = await fetch(getApiURL(this.snipeURL, `/hardware/byserial/${serial}`), {
 			method: "GET",
 			headers: {
@@ -90,8 +84,6 @@ export class HardwareManager extends Manager {
 			}
 		});
 		const result = await res.json();
-		if (res.status !== 200) throw (JSON.stringify(result, null, " "));
-		if (result.status == "error") throw (JSON.stringify(result, null, " "));
 
 		const json: Response<IHardware> = result;
 		return json.rows.map(hardware => new Hardware(hardware));
@@ -105,12 +97,11 @@ export class HardwareManager extends Manager {
 	 * @param location Location ID
 	 */
 	async checkin(id: number, note?: string, location_id?: number) {
+		if(!id) return "id required!";
 		const data = {
 			note: note,
 			location_id: location_id
 		};
-		const JSONdata = JSON.stringify(data);
-		console.log(JSONdata);
 		const res = await fetch(getApiURL(this.snipeURL, `/hardware/${id}/checkin`), {
 			method: "POST",
 			headers: {
@@ -118,13 +109,9 @@ export class HardwareManager extends Manager {
 				"Accept": "application/json",
 				"Content-Type": "application/json"
 			},
-			body: JSONdata
+			body: JSON.stringify(data)
 		});
 		const result = await res.json();
-		if (await result.status == "error") {
-			if (await result.messages == "That asset is already checked in.") throw ("Asset already checked in");
-			throw (`Error on checkin:\n${JSON.stringify(result, null, " ")}`);
-		}
 		return result;
 	}
 
@@ -134,6 +121,7 @@ export class HardwareManager extends Manager {
 	 * @param options Options to pass to the API
 	 */
 	async checkout(id: number, options: checkoutOptions) {
+		if(!id) return "id required!";
 		const res = await fetch(getApiURL(this.snipeURL, `/hardware/${id}/checkout`), {
 			method: "POST",
 			headers: {
@@ -144,10 +132,6 @@ export class HardwareManager extends Manager {
 			body: JSON.stringify(options)
 		});
 		const result = await res.json();
-		if (await result.status == "error") {
-			if (await result.messages == "That asset is not available for checkout!") throw ("Asset already checked out");
-			throw (`Error on checkout:\n${JSON.stringify(result, null, " ")}`);
-		}
 		return result;
 	}
 
@@ -165,8 +149,6 @@ export class HardwareManager extends Manager {
 			}
 		});
 		const result = await res.json();
-		if (res.status !== 200) throw (JSON.stringify(result, null, " "));
-		if (result.status == "error") throw (JSON.stringify(result, null, " "));
 
 		const json: Response<IHardware> = result;
 		return json.rows.map(hardware => new Hardware(hardware));
@@ -185,8 +167,6 @@ export class HardwareManager extends Manager {
 			}
 		});
 		const result = await res.json();
-		if (res.status !== 200) throw (JSON.stringify(result, null, " "));
-		if (result.status == "error") throw (JSON.stringify(result, null, " "));
 
 		const json: Response<IHardware> = result;
 		return json.rows.map(hardware => new Hardware(hardware));
@@ -199,6 +179,7 @@ export class HardwareManager extends Manager {
 	 * @param id Asset ID
 	 */
 	async delete(id: number) {
+		if(!id) return "id required!";
 		const res = await fetch(getApiURL(this.snipeURL, `/hardware/${id}`), {
 			method: "DELETE",
 			headers: {
@@ -208,8 +189,6 @@ export class HardwareManager extends Manager {
 			}
 		});
 		const result = await res.json();
-		if (res.status !== 200) throw (JSON.stringify(result, null, " "));
-		if (result.status == "error") throw (JSON.stringify(result, null, " "));
 
 		return `Asset ${id} deleted`;
 	}
@@ -220,6 +199,7 @@ export class HardwareManager extends Manager {
 	 * Make new asset
 	 */
 	async new(options: HardwarePostOptions) {
+		if(!options.asset_tag || !options.model_id || !options.status_id) return "missing fields!";
 		const res = await fetch(getApiURL(this.snipeURL, "/hardware"), {
 			method: "POST",
 			headers: {
@@ -230,8 +210,6 @@ export class HardwareManager extends Manager {
 			body: JSON.stringify(options)
 		});
 		const result = await res.json();
-		if (res.status !== 200) throw (JSON.stringify(result, null, " "));
-		if (result.status == "error") throw (JSON.stringify(result, null, " "));
 		if (result.status == "success") {
 			const json: IHardware = result.payload;
 			return new Hardware(json);
@@ -245,6 +223,7 @@ export class HardwareManager extends Manager {
 	 * @param options Options to pass to the API
 	 */
 	async update(id: number, options: HardwareUpdateOptions) {
+		if(!id) return "id required!";
 		const res = await fetch(getApiURL(this.snipeURL, `/hardware/${id}`), {
 			method: "PATCH",
 			headers: {
@@ -255,13 +234,31 @@ export class HardwareManager extends Manager {
 			body: JSON.stringify(options)
 		});
 		const result = await res.json();
-		if (res.status !== 200) throw (JSON.stringify(result, null, " "));
-		if (result.status == "error") throw (JSON.stringify(result, null, " "));
 		if (result.status == "success") {
 			const json: IHardware = result.payload;
 			return new Hardware(json);
 		} else {
 			return result;
 		}
+	}
+
+	async audit(asset_tag: string, location_id?: number) {
+		if(!asset_tag) return "asset_tag required!";
+
+		const data = {
+			asset_tag: asset_tag,
+			location_id: location_id
+		};
+		const res = await fetch(getApiURL(this.snipeURL, "/hardware/audit"), {
+			method: "POST",
+			headers: {
+				"Authorization": `Bearer ${this.apiToken}`,
+				"Accept": "application/json",
+				"Content-Type": "application/json"
+			},
+			body: JSON.stringify(data)
+		});
+		const result = await res.json();
+		return result;
 	}
 }
